+++ extras/harmonics_w5/README.md
@@ -1,0 +1,9 @@
+# Harmonics w5
+
+This is a standalone package that provides analytical continuation of weight 5 harmonics.
+
+to install it run:
+
+```
+poetry install
+```
